backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
defusedxml==0.7.1
Django==4.2.16
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-extra-fields==3.7.0
environs==14.6.0
filetype==1.2.0
gunicorn==25.1.0
idna==3.11
marshmallow==4.2.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.8.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
